Move tests to be colocated
